move code to separate array test file and fake classes file
This is simply a move to a more organized file structure. No code was
changed. The "test_fakes" file contains the sample classes used in the
tests.
